[memo-category-dialog] refactor : category 관련 작업들 category 패키지로 묶음
