--- feature/src/main/java/com/hegunhee/newsimplememoapp/feature/common/CategoryActionHandler.kt
@@ -1,10 +1,0 @@
-package com.hegunhee.newsimplememoapp.feature.common
-
-interface CategoryActionHandler {
-
-    fun onBottomSheetDismiss()
-
-    fun onCategoryAdd(categoryType : CategoryType)
-
-    fun onCategoryClick(type : CategoryType,category : String)
-}
