+++ feature/src/main/java/com/hegunhee/newsimplememoapp/feature/common/category/CategoryActionHandler.kt
@@ -1,0 +1,10 @@
+package com.hegunhee.newsimplememoapp.feature.common.category
+
+interface CategoryActionHandler {
+
+    fun onBottomSheetDismiss()
+
+    fun onCategoryAdd(categoryType : CategoryType)
+
+    fun onCategoryClick(type : CategoryType, category : String)
+}
