+++ feature/src/main/java/com/hegunhee/newsimplememoapp/feature/common/category/CategoryType.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.newsimplememoapp.feature.common
+package com.hegunhee.newsimplememoapp.feature.common.category
 
 sealed interface CategoryType {
