+++ feature/src/main/res/layout/include_bottom_category.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="categoryType"
-            type="com.hegunhee.newsimplememoapp.feature.common.CategoryType" />
+            type="com.hegunhee.newsimplememoapp.feature.common.category.CategoryType" />
 
         <variable
             name="categoryList"
@@ -19,7 +19,7 @@
 
         <variable
             name="actionHandler"
-            type="com.hegunhee.newsimplememoapp.feature.common.CategoryActionHandler" />
+            type="com.hegunhee.newsimplememoapp.feature.common.category.CategoryActionHandler" />
 
     </data>
